Moved MainMenu to Application
